# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DummyJSON API",
    "version": "0.0.5",
    "description": "DummyJSON API - Products and Posts endpoints"
  },
...
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "summary": "refresh auth session",
        "tags": ["Auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  },
                  "expiresInMins": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
...
    }
  },
...
}